+++ conf/data/registration/eth2_base.yaml
@@ -32,15 +32,6 @@ data:
     - transform: ComposeTransform
       params:
         transform_options:
-          - transform: IrregularSampling
-            params:
-              d_half: 3.5
-              p: 1
-              grid_size_center: 0.01
-
-    - transform: ComposeTransform
-      params:
-        transform_options:
           - transform: CubeCrop
             params:
               c: 5.5
